+++ kite/src/main/kotlin/com/tang/kite/enumeration/transaction/TransactionIsolationLevel.kt
@@ -1,4 +1,4 @@
-package com.tang.kite.session
+package com.tang.kite.enumeration.transaction
 
 import java.sql.Connection
 
@@ -8,8 +8,8 @@ import java.sql.Connection
 enum class TransactionIsolationLevel(val level: Int) {
 
     NONE(Connection.TRANSACTION_NONE),
-    READ_COMMITTED(Connection.TRANSACTION_READ_COMMITTED),
     READ_UNCOMMITTED(Connection.TRANSACTION_READ_UNCOMMITTED),
+    READ_COMMITTED(Connection.TRANSACTION_READ_COMMITTED),
     REPEATABLE_READ(Connection.TRANSACTION_REPEATABLE_READ),
     SERIALIZABLE(Connection.TRANSACTION_SERIALIZABLE)